+++ src/main/java/net/seanomik/energeticstorage/utils/Utils.java
@@ -72,15 +72,15 @@ public class Utils {
         return false;
     }
 
-    public static Map<ItemStack, Integer> removeSimilarItem(Map<ItemStack, Integer> itemStacks, ItemStack item) {
+    public static Map<ItemStack, Long> removeSimilarItem(Map<ItemStack, Long> itemStacks, ItemStack item) {
         removeAmountFromLore(item);
         // Doing this does not work. So we're gonna have to make some ugly code...
         // The reason it doesn't work is due to Spigot implementing its own `hashCode` implementation,
         // which Java's Iterator#remove method relies on and regenerates the hashCode when removing.
         //itemStacks.entrySet().removeIf(entry -> removeAmountFromLore(entry.getKey()).isSimilar(item));
 
-        Map<ItemStack, Integer> items = new HashMap<>();
-        for (Map.Entry<ItemStack, Integer> entry : itemStacks.entrySet()) {
+        Map<ItemStack, Long> items = new HashMap<>();
+        for (Map.Entry<ItemStack, Long> entry : itemStacks.entrySet()) {
             if (!removeAmountFromLore(entry.getKey()).isSimilar(item)) {
                 items.put(entry.getKey(), entry.getValue());
             }
